Simplification of Project Structure / Dealing with Parentheses
We can now run our scrapy script and web spider from within the same Python script!  Although, we'll eventually have the web spider definition and the running of our application in different files, since the parse function is already getting pretty long.

Also, now we search for something of the form "/wiki/something_like_this_(Something)/" in our <p> tags, and check to see if what precedes this link is balanced with respect to parentheses.  If it isn't, continue to the next link in the paragraph, and if there isn't one, then proceed to the next paragraph.  This works for Hank Green (we get to philosophy) and Tom Hanks (we get trapped in Semiotics <=> Meaning (Semiotics))!

However, there are a couple issues at hand.  We click on the wrong link on Germany because there's some Coordinates at the top of the article in a span, containing a hyperlink in a <p> tag (although we should be able to ignore <p> tags if they contain this span id?).  Also, I'm not sure how we're handing italicized links, but I should definitely give more thought to that.

Maybe this should just be a Python Game instead of a Javascript-y game?  The problem is that my parents and friends will have a higher barrier to entry.  User experience is always best on integrated websites.  Also I'm afraid of the unknown and I have no idea how to execute server-side scripts or set any of that stuff up or what I even need to do in order to deliver that functionality.  But I should definitely play around with it, because it's cool.