Coordinates and Non-standard Characters Fixed
Our parser now checks that a paragraph DOES NOT contain the span tag for coordinates before attempting to process it (otherwise, it just moves onto the next paragraph).

Also, we now write to the list_of_pages.txt file in binary mode and UTF-8 encoding, so that we can use an expanded character set.  This is necessary for words like "Maori", which we got from Wikipedia's random article feature!  I love Wikipedia, man.

On this point, we also updated our regex pattern, so we're matching the link to the terminal end quote ("), and also allowing the matching of - and % and # to deal with hyperlinks like "Proto-Indo-European" and "Urban_M%C4%81ori" and "Internalization#Psychology_and_sociology".

Also we now terminate at "Philosophy" as required.  Note that the spider automatically terminates when requesting a duplicate page, from the native scrapy handling of Spiders and Requests.